+++ src/pytezos/cli/cli.py
@@ -31,9 +31,7 @@ from pytezos.sandbox.node import DOCKER_IMAGE
 from pytezos.sandbox.node import TEZOS_NODE_PORT
 from pytezos.sandbox.node import SandboxedNodeContainer
 from pytezos.sandbox.node import get_next_baker_key
-from pytezos.sandbox.parameters import HANGZHOU
-from pytezos.sandbox.parameters import ITHACA
-from pytezos.sandbox.parameters import JAKARTA
+from pytezos.sandbox.parameters import KATHMANDU
 
 kernel_js_path = join(dirname(dirname(__file__)), 'assets', 'kernel.js')
 kernel_json = {
@@ -342,7 +340,7 @@ def smartpy_compile(
 
 @cli.command(help='Run containerized sandbox node')
 @click.option('--image', type=str, help='Docker image to use', default=DOCKER_IMAGE)
-@click.option('--protocol', type=click.Choice(['jakarta']), help='Protocol to use', default='jakarta')
+@click.option('--protocol', type=click.Choice(['kathmandu']), help='Protocol to use', default='kathmandu')
 @click.option('--port', '-p', type=int, help='Port to expose', default=TEZOS_NODE_PORT)
 @click.option('--interval', '-i', type=float, help='Interval between baked blocks (in seconds)', default=1.0)
 @click.option('--blocks', '-b', type=int, help='Number of blocks to bake before exit')
@@ -356,7 +354,7 @@ def sandbox(
     blocks: int,
 ):
     protocol_hash = {
-        'jakarta': JAKARTA,
+        'kathmandu': KATHMANDU,
     }[protocol]
 
     node = SandboxedNodeContainer(image=image, port=port)