+++ src/pytezos/michelson/tags.py
@@ -154,6 +154,8 @@ prim_tags = {
     'tx_rollup_l2_address': b'\x94',
     'MIN_BLOCK_TIME': b'\x95',
     'sapling_transaction': b'\x96',
+    # KATHMANDU
+    'EMIT': b'\x97',
     # FIXME: Dummy values for TZT, refactor macros
     'Stack_elt': b'\xEE',
     'Big_map': b'\xEE',