+++ src/pytezos/sandbox/node.py
@@ -22,7 +22,7 @@ from pytezos.operation.group import OperationGroup
 from pytezos.sandbox.parameters import LATEST
 from pytezos.sandbox.parameters import sandbox_addresses
 
-DOCKER_IMAGE = 'bakingbad/sandboxed-node:v13.0-1'
+DOCKER_IMAGE = 'bakingbad/sandboxed-node:v14.0-1'
 MAX_ATTEMPTS = 100
 ATTEMPT_DELAY = 0.1
 TEZOS_NODE_PORT = 8732